+++ src/api/viewsets/dataset/view.py
@@ -11,6 +11,8 @@ class DatasetViewSet(viewsets.ModelViewSet):
     queryset = Dataset.objects.all()
     serializer_class = DatasetSerializer
 
+    dataset_service = DatasetService()
+
     def list(self, request, *args, **kwargs):
         if request.user.role == 'ADMIN' and request.user.company:
             users = RegisterUserService().get_users_by_company_id(request.user.company_id)
@@ -21,8 +23,12 @@ class DatasetViewSet(viewsets.ModelViewSet):
         json_datasets = DatasetSerializer(datasets, many=True).data
         return Response(json_datasets, status=status.HTTP_200_OK)
 
-    @staticmethod
-    def create(request, *args, **kwargs):
+    def create(self, request, *args, **kwargs):
         dataset = {'name': request.data['name'], 'user': request.user.uuid, 'project': request.data['project']}
-        created_dataset = DatasetService().create_dataset(dataset)
+        created_dataset = self.dataset_service.create(dataset)
         return Response(created_dataset, status=status.HTTP_201_CREATED)
+
+    def destroy(self, request, pk=None, *args, **kwargs):
+        destroyed = self.dataset_service.delete(pk)
+        return Response(status=status.HTTP_204_NO_CONTENT) if destroyed > 0 \
+            else Response("Resource not found", status=status.HTTP_404_NOT_FOUND)