+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file
 ## [0.0.8](https://github.com/pablobascunana/youml-manager/compare/90712cb...develop) (WIP)
 
 * [ML-68](https://youml.atlassian.net/browse/ML-68) Delete project
+* [ML-69](https://youml.atlassian.net/browse/ML-69) Delete dataset and rename tests cases
 
 
 ## [0.0.7](https://github.com/pablobascunana/youml-manager/compare/cd3a3b9...90712cb) (16-10-2022)